pyproject.toml
[project]
name = "blog-generationapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.135.3",
    "pydantic[email]>=2.12.5",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]==1.7.4",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.43.0",
    "psycopg[binary]>=3.3.3",
    "bcrypt==3.2.2",
    "python-dotenv>=1.2.2",
]
